Enhance UI/UX with Material 3 design and fix compilation errors
- Implement Display Mode and Edit Mode in AddEditDetailView with spring-based animations
- Add sophisticated fade and slide transitions between modes
- Integrate Add Funds dialog with savings goal progress tracking
- Fix compilation errors: stringResource calls in coroutines and type casting
- Enhance WishListScreen with interactive TodosContent composable
- Integrate unused composables: EnhancedEmptyWishListState, EmptySearchState
- Add haptic feedback and spring animations to interactive elements
- Implement proper Material 3 styling throughout the app
- Add comprehensive string resources and centralized styling
- Fix scope issues and parameter passing in navigation callbacks
- Integrate GoalProgressBar with proper Float type handling
- Add WishViewModelFactory for proper dependency injection
- Update build configuration and dependencies for enhanced features

+++ app/src/main/java/com/example/mywishlistapp/GoalProgressBar.kt
@@ -8,18 +8,15 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import kotlin.math.min
 
 @Composable
 fun GoalProgressBar(
     title: String,
-    currentAmount: Double,
-    targetAmount: Double,
+    currentAmount: Float,
+    targetAmount: Float,
     currency: String = "$",
     modifier: Modifier = Modifier
 ) {
@@ -118,8 +115,8 @@ fun SavingsGoalSection(
         
         GoalProgressBar(
             title = "Total Savings Goal",
-            currentAmount = currentSavings,
-            targetAmount = savingsGoal,
+            currentAmount = currentSavings.toFloat(),
+            targetAmount = savingsGoal.toFloat(),
             currency = "$"
         )
     }
